+++ include/crab/http/request_parser.hpp
@@ -89,7 +89,7 @@ struct BodyParser {
 
 public:
 	BodyParser() = default;
-	BodyParser(details::optional<uint64_t> content_length, bool chunked);
+	BodyParser(optional<uint64_t> content_length, bool chunked);
 
 	StringStream body;
 	size_t max_chunk_header_total_length = 256;